// src/app/common/interfaces/user.ts
export interface IUser {
    id: string;
    name: string;
    email: string;
}

export interface IHome {
    frontCover: string;
    frontCover2: string;
    frontText1: string;
    frontText2: string;
    frontText3: string;
    frontText4: string;
    frontText5: string;
    frontText6: string;
    frontText7: string;
    frontText8: string;
    bottomCover: string;
    bottomText: string;
}